requirements.txt
python-dotenv==1.0.1
requests==2.31.0
pip install python-telegram-bot==13.0
